splt all dungeon cutscenes files
