--- data/data_8107224.s
@@ -1,13 +1,0 @@
-.section .rodata
-
-
-
-.string "pksdir0\0" @Mewtwo
-.string "pksdir0\0" @Entei
-.string "pksdir0\0" @Raikou
-.string "pksdir0\0" @Suicune
-.string "pksdir0\0" @HoOh
-.string "pksdir0\0" @Latios
-.string "pksdir0\0" @ regis
-
-.align 2,0
