+++ ld_script.txt
@@ -292,7 +292,13 @@ SECTIONS {
         src/dungeon_cutscene_magma_cavern.o(.text);
         src/dungeon_cutscene_rayquaza.o(.text);
         src/dungeon_cutscene_mankey.o(.text);
-        src/dungeon_cutscenes.o(.text);
+        src/dungeon_cutscene_mewtwo.o(.text);
+        src/dungeon_cutscene_entei.o(.text);
+        src/dungeon_cutscene_raikou.o(.text);
+        src/dungeon_cutscene_suicune.o(.text);
+        src/dungeon_cutscene_hooh.o(.text);
+        src/dungeon_cutscene_latios.o(.text);
+        src/dungeon_cutscene_regis.o(.text);
         src/dungeon_cutscene_jirachi.o(.text);
         src/dungeon_cutscene_lugia.o(.text);
         src/dungeon_cutscene_kyogre.o(.text);
@@ -651,7 +657,13 @@ SECTIONS {
         src/dungeon_cutscene_magma_cavern.o(.rodata);
         src/dungeon_cutscene_rayquaza.o(.rodata);
         src/dungeon_cutscene_mankey.o(.rodata);
-        data/data_8107224.o(.rodata);
+        src/dungeon_cutscene_mewtwo.o(.rodata);
+        src/dungeon_cutscene_entei.o(.rodata);
+        src/dungeon_cutscene_raikou.o(.rodata);
+        src/dungeon_cutscene_suicune.o(.rodata);
+        src/dungeon_cutscene_hooh.o(.rodata);
+        src/dungeon_cutscene_latios.o(.rodata);
+        src/dungeon_cutscene_regis.o(.rodata);
         src/dungeon_cutscene_jirachi.o(.rodata);
         src/dungeon_cutscene_lugia.o(.rodata);
         src/dungeon_cutscene_kyogre.o(.rodata);
