+++ src/dungeon_cutscene_mewtwo.c
@@ -1,0 +1,221 @@
+#include "global.h"
+#include "globaldata.h"
+#include "code_800DAC0.h"
+#include "code_800E9A8.h"
+#include "code_803E46C.h"
+#include "code_803E668.h"
+#include "code_8041AD0.h"
+#include "code_804267C.h"
+#include "code_80861A8.h"
+#include "code_806CD90.h"
+#include "code_80869E4.h"
+#include "constants/bg_music.h"
+#include "constants/direction.h"
+#include "constants/friend_area.h"
+#include "constants/weather.h"
+#include "dungeon_items.h"
+#include "dungeon_leader.h"
+#include "dungeon_map_access.h"
+#include "dungeon_message.h"
+#include "dungeon_music.h"
+#include "dungeon_misc.h"
+#include "dungeon_logic.h"
+#include "dungeon_random.h"
+#include "dungeon_util_1.h"
+#include "dungeon_util.h"
+#include "exclusive_pokemon.h"
+#include "friend_area.h"
+#include "items.h"
+#include "pokemon_3.h"
+#include "pokemon.h"
+#include "position_util.h"
+#include "trap.h"
+#include "math.h"
+#include "dungeon_config.h"
+#include "dungeon_boss_dialogue.h"
+
+extern void sub_8042B0C(Entity *);
+extern u8 sub_806FD18(Entity *);
+extern void sub_806FDF4(Entity *, Entity *, Entity **);
+extern s32 GetCameraXPos(void);
+extern s32 GetCameraYPos(void);
+extern void sub_803F878(u32, s32);
+extern void BgColorCallNullsub4(void);
+extern void sub_8085EB0(void);
+extern void sub_8049884(void);
+extern void sub_8049B8C(void);
+extern void ShowWholeRevealedDungeonMap(void);
+extern void sub_8086A54(Entity *);
+extern void ResetMonEntityData(EntityInfo *, u32);
+extern void sub_8041888(u32);
+extern u32 sub_80861F8(u32, Entity *, u32);
+extern u8 sub_80860A8(u32);
+extern void sub_8052D44(s16 *, Entity *, Entity *);
+extern void sub_80464C8(Entity *, DungeonPos *, Item *);
+extern void SetDungeonBGColorRGB(u32, u32, u32, u32, u32);
+extern u32 sub_8085EC8(u32, u32, u32, DungeonPos *, u32);
+extern void sub_807EAA0(u32, u32);
+extern void sub_8072008(Entity *, Entity *, s16, u32, u32);
+extern void sub_8085374(void);
+extern u32 GetRandomFloorItem(u32);
+extern bool8 sub_8085B80(struct_8085B80 *);
+extern void sub_8086A3C(Entity *pokemon);
+extern void SetupBossFightHP(Entity *pokemon, s32 newHP, u16 songIndex);
+extern u8 sub_8086AE4(s16 _index);
+
+static void MewtwoDropInEffect(Entity *mewtwoEntity);
+static void MewtwoScreenFlash(void);
+
+void sub_8089914(void)
+{
+  Entity *leaderEntity;
+  Entity *MewtwoEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  MewtwoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MEWTWO);
+  DungeonStopBGM();
+  gDungeon->unk7 = 1;
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  sub_80855E4(sub_8086A3C);
+  SetFacingDirection(MewtwoEntity, DIRECTION_SOUTH);
+  sub_8086A3C(MewtwoEntity);
+  sub_8085860(leaderEntity->pos.x,leaderEntity->pos.y - 3);
+  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2],MONSTER_MEWTWO);
+}
+
+void sub_8089978(void)
+{
+  Entity *leaderEntity;
+  Entity *MewtwoEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  MewtwoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MEWTWO);
+  DungeonStopBGM();
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  sub_80855E4(sub_8086A3C);
+  if(HasRecruitedMon(MONSTER_MEWTWO)){
+    DungeonStartNewBGM(MUS_FRIEND_AREA_CRYPTIC_CAVE);
+    sub_8068FE0(MewtwoEntity,0x21c,0);
+  }
+  else {
+    gDungeon->unk7 = 1;
+    SetFacingDirection(MewtwoEntity, DIRECTION_SOUTH);
+    sub_8086A3C(MewtwoEntity);
+  }
+  sub_8085860(leaderEntity->pos.x,leaderEntity->pos.y - 3);
+  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2],MONSTER_MEWTWO);
+}
+
+void sub_8089A00(u8 param_1, u8 param_2)
+{
+  if ((param_2 == 0x1d || param_2 == 0x1e || param_2 == 0x1f) && (param_1 == 0x11)) {
+    sub_8097FA8(0x13);
+    gDungeon->unk2 = 1;
+  }
+}
+
+void MewtwoPreFightDialogue(void)
+{
+  Entity *leaderEntity;
+  Entity *MewtwoEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  MewtwoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MEWTWO);
+  sub_8086448();
+  DisplayDungeonDialogue(&gMewtwoPreFightDialogue_1);
+  sub_803E708(10,0x46);
+  DungeonStartNewBGM(MUS_FRIEND_AREA_CRYPTIC_CAVE);
+  sub_8086794();
+  sub_803E708(0x1e,0x46);
+  MewtwoDropInEffect(MewtwoEntity);
+  sub_803E708(0x3c,0x46);
+  DisplayDungeonDialogue(&gMewtwoPreFightDialogue_2);
+  DisplayDungeonDialogue(&gMewtwoPreFightDialogue_3);
+  sub_803E708(10,0x46);
+  MewtwoScreenFlash();
+  DisplayDungeonDialogue(&gMewtwoPreFightDialogue_4);
+  MewtwoScreenFlash();
+  DisplayDungeonDialogue(&gMewtwoPreFightDialogue_5);
+  sub_803E708(10,0x46);
+  SetupBossFightHP(MewtwoEntity,900,MUS_BATTLE_WITH_RAYQUAZA);
+  ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
+}
+
+void MewtwoReFightDialogue(void)
+{
+  Entity *leaderEntity;
+  Entity *MewtwoEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  MewtwoEntity = GetEntityFromMonsterBehavior(BEHAVIOR_MEWTWO);
+  sub_8086448();
+  if(HasRecruitedMon(MONSTER_MEWTWO)) {
+    sub_80866C4(&gMewtwoReFightDialogue_5);
+  }
+  else {
+    DisplayDungeonDialogue(&gMewtwoReFightDialogue_1);
+    DungeonStartNewBGM(MUS_FRIEND_AREA_CRYPTIC_CAVE);
+    sub_8086794();
+    MewtwoDropInEffect(MewtwoEntity);
+    sub_803E708(0x3c,0x46);
+    DisplayDungeonDialogue(&gMewtwoReFightDialogue_2);
+    MewtwoScreenFlash();
+    DisplayDungeonDialogue(&gMewtwoReFightDialogue_3);
+    MewtwoScreenFlash();
+    DisplayDungeonDialogue(&gMewtwoReFightDialogue_4);
+    SetupBossFightHP(MewtwoEntity,900,MUS_BATTLE_WITH_RAYQUAZA);
+    ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
+  }
+}
+
+static void MewtwoDropInEffect(Entity *mewtwoEntity)
+{
+  s24_8 iVar1;
+  s24_8 iVar2;
+
+  GetEntInfo(mewtwoEntity)->unk15E = 0;
+  iVar2 = IntToF248_2(200);
+  iVar1 = IntToF248_2(4);
+  PlaySoundEffect(0x1f8);
+  while( 1 ) {
+    iVar2.raw -= iVar1.raw;
+    iVar1 = F248_Sub(iVar1, FloatToF248_2(0.045));
+    if (F248LessThanFloat(iVar1, 0.12)) {
+      iVar1 = FloatToF248_2(0.12);
+    }
+    if (F248LessThanInt(iVar2, 0)) break;
+    GetEntInfo(mewtwoEntity)->unk174 = iVar2;
+    DungeonRunFrameActions(0x46);
+  }
+  GetEntInfo(mewtwoEntity)->unk174 = IntToF248_2(0);
+}
+
+static void MewtwoScreenFlash(void)
+{
+  int iVar1;
+
+  PlaySoundEffect(0x1f6);
+  for(iVar1 = 250; iVar1 > 149; iVar1 -= 10)
+  {
+    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,1,1);
+    DungeonRunFrameActions(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 > 199; iVar1 -= 10)
+  {
+    SetDungeonBGColorRGB(iVar1,iVar1,0,1,1);
+    DungeonRunFrameActions(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 10)
+  {
+    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,1,1);
+    DungeonRunFrameActions(0x46);
+  }
+  sub_803E708(10,0x46);
+  sub_8085EB0();
+}
+
+
