--- src/dungeon_cutscene_placeholder.c
@@ -1,65 +1,0 @@
-#include "global.h"
-#include "globaldata.h"
-#include "code_800DAC0.h"
-#include "code_800E9A8.h"
-#include "code_803E46C.h"
-#include "code_803E668.h"
-#include "code_8041AD0.h"
-#include "code_804267C.h"
-#include "code_80861A8.h"
-#include "code_806CD90.h"
-#include "code_80869E4.h"
-#include "constants/bg_music.h"
-#include "constants/direction.h"
-#include "constants/friend_area.h"
-#include "constants/weather.h"
-#include "dungeon_items.h"
-#include "dungeon_leader.h"
-#include "dungeon_map_access.h"
-#include "dungeon_message.h"
-#include "dungeon_music.h"
-#include "dungeon_misc.h"
-#include "dungeon_logic.h"
-#include "dungeon_random.h"
-#include "dungeon_util_1.h"
-#include "dungeon_util.h"
-#include "exclusive_pokemon.h"
-#include "friend_area.h"
-#include "items.h"
-#include "pokemon_3.h"
-#include "pokemon.h"
-#include "position_util.h"
-#include "trap.h"
-#include "math.h"
-#include "dungeon_config.h"
-#include "dungeon_boss_dialogue.h"
-
-extern void sub_8042B0C(Entity *);
-extern u8 sub_806FD18(Entity *);
-extern void sub_806FDF4(Entity *, Entity *, Entity **);
-extern s32 GetCameraXPos(void);
-extern s32 GetCameraYPos(void);
-extern void sub_803F878(u32, s32);
-extern void BgColorCallNullsub4(void);
-extern void sub_8085EB0(void);
-extern void sub_8049884(void);
-extern void sub_8049B8C(void);
-extern void ShowWholeRevealedDungeonMap(void);
-extern void sub_8086A54(Entity *);
-extern void ResetMonEntityData(EntityInfo *, u32);
-extern void sub_8041888(u32);
-extern u32 sub_80861F8(u32, Entity *, u32);
-extern u8 sub_80860A8(u32);
-extern void sub_8052D44(s16 *, Entity *, Entity *);
-extern void sub_80464C8(Entity *, DungeonPos *, Item *);
-extern void SetDungeonBGColorRGB(u32, u32, u32, u32, u32);
-extern u32 sub_8085EC8(u32, u32, u32, DungeonPos *, u32);
-extern void sub_807EAA0(u32, u32);
-extern void sub_8072008(Entity *, Entity *, s16, u32, u32);
-extern void sub_8085374(void);
-extern u32 GetRandomFloorItem(u32);
-extern bool8 sub_8085B80(struct_8085B80 *);
-extern void sub_8086A3C(Entity *pokemon);
-extern void SetupBossFightHP(Entity *pokemon, s32 newHP, u16 songIndex);
-extern u8 sub_8086AE4(s16 _index);
-
